refactor: WI-0653 extract scheduling rotation fairness core helpers
