+++ ROADMAP.md
@@ -934,3 +934,4 @@ Phase 8: Extensions (ATS, performance, expenses, analytics)
 - WI-0650 employee attendance prefill effect extraction + line-budget recovery (extract schedule-context correction prefill useEffect from /employee/page.tsx into page-attendance-prefill-effect hook, preserve auto-target selection behavior, and keep employee/page.tsx <=500 lines with e2e-wi0650 guard)
 - WI-0651 admin approval-executions section file split (split page-sections.tsx into work-conditions / summary-escalation / queue section modules, keep each file <=300 lines, and add e2e-wi0651 split+line-budget guard)
 - WI-0652 korean surface mojibake guard and people copy fix (fix corrupted ko copy in /admin/people title/action/kpi/log labels and add e2e-wi0652 guard for admin-people + withholding runtime labels + contracts runtime fallback labels)
+- WI-0653 scheduling rotation fairness core helper extraction (extract template/employee normalization, global/advanced fairness constraint normalization, planned-minute/grade evaluators, and weekday-set helper from scheduling/service.ts into rotation-fairness-core-helpers.ts, keep behavior unchanged, and add e2e-wi0653 regression guard)
